comments typings fix + bookmark stringify fix
