+++ portal/src/views/comments/model.ts
@@ -1,0 +1,16 @@
+export interface Comment {
+    id: number;
+    author: object;
+    bookmark?: string;
+    body: string;
+    replies: Comment[];
+    createdAt: number;
+    updatedAt: number;
+}
+
+export interface NewComment {
+    projectId: number;
+    bookmark: string;
+    body: string;
+    threadId?: number;
+}
